requirements.txt
pillow>=10.4.0
numpy>=1.23.0
pandas>=2.0.1
pyautogui>=0.9.54
pynput>=1.8.1
